Moved RubySL::Socket::Helpers into RubySL::Socket
